lesson-07-rag/exercises/starter/shopping/requirements.txt
google-adk>=1.17.0
toolbox-core>=0.5.0
google-cloud-discoveryengine>=0.11.0